+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # v0.5
 - Added functionality to import tokens from Google Authenticator migration QR codes.
 - UI Refactor
-- 
+- Fixed system tray icon size
 
 # v0.4
 - Added next code preview functionality